JUMPY_TEST_SUITE, change code for test discovery.
